requirements.txt
asgiref==3.4.1
Django==3.2.5
django-environ==0.4.5
django-filter==2.4.0
djangorestframework==3.12.4
Markdown==3.3.4
pytz==2021.1
sqlparse==0.4.1
